create login jsp and index jsp pages
login.jsp: With this page, users can fill in their username and password on a sign-in form. A link to the registration page for new users and an image are also provided .

index.jsp: This page checks if a user is logged in by verifying the presence of a "name" attribute in the session. If the user is not logged in, it redirects them to the login page.

create the fonts and images folders and added css .